requirements.txt
pdfminer.six
requests
colorama
